# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RomaniaDate Dating App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /swipe/like:
    post:
      operationId: likeUser
      tags: [swipe]
      summary: Like a user (costs 10 diamonds)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeRequest"
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResponse"
        "402":
          description: Not enough diamonds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /swipe/dislike:
    post:
      operationId: dislikeUser
      tags: [swipe]
      summary: Dislike a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeRequest"
      responses:
        "200":
          description: Disliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResponse"
  /swipe/superlike:
    post:
      operationId: superlikeUser
      tags: [swipe]
      summary: Superlike a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeRequest"
      responses:
        "200":
          description: Superliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResponse"
        "429":
          description: Daily limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /swipe/rewind:
    post:
      operationId: rewindSwipe
      tags: [swipe]
      summary: Rewind previous swipe (costs 30 diamonds)
      responses:
        "200":
          description: Rewound
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResponse"
        "402":
          description: Not enough diamonds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SwipeRequest:
      type: object
      properties:
        targetUserId:
          type: integer
      required:
        - targetUserId
    SwipeResponse:
      type: object
      properties:
        success:
          type: boolean
        isMatch:
          type: boolean
        matchId:
          type: ["integer", "null"]
        diamondsSpent:
          type: integer
        diamondsRemaining:
          type: integer
        message:
          type: ["string", "null"]
      required:
        - success
        - isMatch
        - diamondsSpent
        - diamondsRemaining
...